Add rescaling on the image array
This commit makes it so that the image array will have it's RGB
value rescaled from 0-255 to 0.0 to 1.0 in order to reduce the
amount of memory used and increase accuracy whenever generating the
model. It is done after looping through all images and creating the
image array. In order to have this working, we imported layers from
keras.

Change the csv paths so that they will be relative to the current
working path. This is a crutch put in place as we don't have a
parser yet.