Initial checkpoint work.
We are now enabling the generation of a model checkpoint between
every and each epoch, and saving them (instead of just preserving
the best results). I'm doing this for now as we don't have a flag
to enable/disable it and I'm not really sure what could be done
by using the checkpoint in it's entirety.